chore: split permission controller
