+++ packages/sheet-permission-ui/src/controller/sheet-permission-interceptor-base.controller.ts
@@ -18,28 +18,25 @@
 /* eslint-disable max-lines-per-function */
 
 import type { ICellData, ICellDataForSheetInterceptor, ICommandInfo, IObjectMatrixPrimitiveType, IPermissionTypes, IRange, ISheetDataValidationRule, Nullable, Workbook, Worksheet } from '@univerjs/core';
-import { DisposableCollection, FOCUSING_EDITOR_STANDALONE, ICommandService, IContextService, IPermissionService, isICellData, IUniverInstanceService, LifecycleStages, LocaleService, ObjectMatrix, OnLifecycle, Rectangle, RxDisposable, UniverInstanceType } from '@univerjs/core';
+import { DisposableCollection, FOCUSING_EDITOR_STANDALONE, ICommandService, IContextService, IPermissionService, isICellData, IUniverInstanceService, LifecycleStages, LocaleService, ObjectMatrix, OnLifecycle, Rectangle, Disposable, UniverInstanceType, Tools } from '@univerjs/core';
 import type { IMoveColsCommandParams, IMoveRangeCommandParams, IMoveRowsCommandParams, ISetRangeValuesCommandParams, ISetSpecificColsVisibleCommandParams, ISetSpecificRowsVisibleCommandParams, ISetWorksheetNameMutationParams } from '@univerjs/sheets';
 import { ClearSelectionContentCommand, DeleteRangeMoveLeftCommand, DeleteRangeMoveUpCommand, DeltaColumnWidthCommand, DeltaRowHeightCommand, getSheetCommandTarget, InsertRangeMoveDownCommand, InsertRangeMoveRightCommand, MoveColsCommand, MoveRangeCommand, MoveRowsCommand, RangeProtectionPermissionEditPoint, RangeProtectionPermissionViewPoint, RangeProtectionRuleModel, SelectionManagerService, SetBackgroundColorCommand, SetColWidthCommand, SetRangeValuesCommand, SetRowHeightCommand, SetSelectedColsVisibleCommand, SetSelectedRowsVisibleCommand, SetSpecificColsVisibleCommand, SetSpecificRowsVisibleCommand, SetWorksheetNameCommand, SetWorksheetNameMutation, SetWorksheetOrderCommand, SetWorksheetRowIsAutoHeightCommand, SetWorksheetShowCommand, WorkbookCopyPermission, WorkbookEditablePermission, WorkbookManageCollaboratorPermission, WorksheetCopyPermission, WorksheetEditPermission, WorksheetFilterPermission, WorksheetProtectionRuleModel, WorksheetSetCellStylePermission, WorksheetSetCellValuePermission, WorksheetSetColumnStylePermission, WorksheetSetRowStylePermission, WorksheetViewPermission } from '@univerjs/sheets';
-import { Inject, Injector } from '@wendellhu/redi';
+import { Inject } from '@wendellhu/redi';
 import { IDialogService } from '@univerjs/ui';
 import { filter, first } from 'rxjs/operators';
 import type { IRenderContext, IRenderController, SpreadsheetSkeleton } from '@univerjs/engine-render';
 import type { ISheetPasteParams } from '@univerjs/sheets-ui';
-import { ApplyFormatPainterCommand, AutoFillCommand, FormulaEditorController, HeaderMoveRenderController, HeaderResizeRenderController, IAutoFillService, ISelectionRenderService, ISheetClipboardService, SetCellEditVisibleOperation, SetRangeBoldCommand, SetRangeItalicCommand, SetRangeStrickThroughCommand, SetRangeUnderlineCommand, SheetCopyCommand, SheetCutCommand, SheetPasteColWidthCommand, SheetPasteShortKeyCommand, StatusBarController, virtualizeDiscreteRanges } from '@univerjs/sheets-ui';
+import { ApplyFormatPainterCommand, AutoFillCommand, IAutoFillService, SetCellEditVisibleOperation, SetRangeBoldCommand, SetRangeItalicCommand, SetRangeStrickThroughCommand, SetRangeUnderlineCommand, SheetCopyCommand, SheetCutCommand, SheetPasteColWidthCommand, SheetPasteShortKeyCommand } from '@univerjs/sheets-ui';
 import { SheetsFilterService } from '@univerjs/sheets-filter';
 import type { IOpenFilterPanelOperationParams } from '@univerjs/sheets-filter-ui';
 import { OpenFilterPanelOperation, SmartToggleSheetsFilterCommand } from '@univerjs/sheets-filter-ui';
-import { SheetsFindReplaceController } from '@univerjs/sheets-find-replace';
 import { InsertCommand } from '@univerjs/docs';
 import type { IUpdateSheetDataValidationRangeCommandParams } from '@univerjs/sheets-data-validation';
-import { AddSheetDataValidationCommand, DataValidationController, DataValidationFormulaController, UpdateSheetDataValidationRangeCommand } from '@univerjs/sheets-data-validation';
+import { AddSheetDataValidationCommand, UpdateSheetDataValidationRangeCommand } from '@univerjs/sheets-data-validation';
 import type { IAddCfCommandParams } from '@univerjs/sheets-conditional-formatting-ui';
-import { AddCfCommand, ConditionalFormattingClearController } from '@univerjs/sheets-conditional-formatting-ui';
-import type { IConditionalFormattingRuleConfig, IConditionFormattingRule } from '@univerjs/sheets-conditional-formatting';
-import { HeaderFreezeRenderController } from '@univerjs/sheets-ui/controllers/render-controllers/freeze.render-controller.js';
+import { AddCfCommand } from '@univerjs/sheets-conditional-formatting-ui';
 import { UnitAction } from '@univerjs/protocol';
-import { deserializeRangeWithSheet, LexerTreeBuilder, NullValueObject } from '@univerjs/engine-formula';
+import { deserializeRangeWithSheet, IDefinedNamesService, LexerTreeBuilder, operatorToken } from '@univerjs/engine-formula';
 import { UNIVER_SHEET_PERMISSION_ALERT_DIALOG, UNIVER_SHEET_PERMISSION_ALERT_DIALOG_ID } from '../views/error-msg-dialog/interface';
 
 type ICellPermission = Record<UnitAction, boolean> & { ruleId?: string; ranges?: IRange[] };
@@ -47,13 +44,12 @@ type ICheckPermissionCommandParams = IMoveRowsCommandParams | IMoveColsCommandPa
 
 export const SHEET_PERMISSION_PASTE_PLUGIN = 'SHEET_PERMISSION_PASTE_PLUGIN';
 
-@OnLifecycle(LifecycleStages.Steady, SheetPermissionInterceptorRenderController)
-export class SheetPermissionInterceptorRenderController extends RxDisposable implements IRenderController {
+@OnLifecycle(LifecycleStages.Steady, SheetPermissionInterceptorBaseController)
+export class SheetPermissionInterceptorBaseController extends Disposable {
     disposableCollection = new DisposableCollection();
 
     constructor(
         private readonly _context: IRenderContext<Workbook>,
-        @Inject(Injector) private readonly _injector: Injector,
         @ICommandService private readonly _commandService: ICommandService,
         @IUniverInstanceService private readonly _univerInstanceService: IUniverInstanceService,
         @IPermissionService private readonly _permissionService: IPermissionService,
@@ -61,39 +57,18 @@ export class SheetPermissionInterceptorRenderController extends RxDisposable imp
         @IDialogService private readonly _dialogService: IDialogService,
         @Inject(RangeProtectionRuleModel) private _rangeProtectionRuleModel: RangeProtectionRuleModel,
         @Inject(WorksheetProtectionRuleModel) private _worksheetProtectionRuleModel: WorksheetProtectionRuleModel,
-        @Inject(HeaderMoveRenderController) private _headerMoveRenderController: HeaderMoveRenderController,
-        @Inject(HeaderResizeRenderController) private _headerResizeRenderController: HeaderResizeRenderController,
-        @ISelectionRenderService private _selectionRenderService: ISelectionRenderService,
         @IAutoFillService private _autoFillService: IAutoFillService,
         @Inject(SheetsFilterService) private _sheetsFilterService: SheetsFilterService,
-        @Inject(SheetsFindReplaceController) private _sheetsFindReplaceController: SheetsFindReplaceController,
         @Inject(LocaleService) private readonly _localService: LocaleService,
-        @Inject(DataValidationController) private readonly _dataValidationController: DataValidationController,
-        @Inject(ConditionalFormattingClearController) private readonly _conditionalFormattingClearController: ConditionalFormattingClearController,
-        @Inject(ISheetClipboardService) private _sheetClipboardService: ISheetClipboardService,
-        @Inject(HeaderFreezeRenderController) private _headerFreezeRenderController: HeaderFreezeRenderController,
         @Inject(LexerTreeBuilder) private readonly _lexerTreeBuilder: LexerTreeBuilder,
         @IContextService private readonly _contextService: IContextService,
-        @Inject(StatusBarController) private readonly _statusBarController: StatusBarController,
-        @Inject(DataValidationFormulaController) private readonly _dataValidationFormulaController: DataValidationFormulaController
+        @IDefinedNamesService private readonly _definedNamesService: IDefinedNamesService
     ) {
         super();
         this._initialize();
-        this._initHeaderMovePermissionInterceptor();
-        this._initHeaderResizePermissionInterceptor();
-        this._initRangeFillPermissionInterceptor();
-        this._initRangeMovePermissionInterceptor();
-        this._initSheetFindPermissionInterceptor();
-        this._initDataValidationPermissionInterceptor();
-        this._initConditionalFormattingPermissionInterceptor();
-        this._initFreezePermissionInterceptor();
-        this._initFormulaEditorPermissionInterceptor();
-        this._initStatusBarPermissionInterceptor();
-        this._initDvFormulaPermissionInterceptor();
-        this._initClipboardHook();
     }
 
-    private _haveNotPermissionHandle(errorMsg: string) {
+    public haveNotPermissionHandle(errorMsg: string) {
         const dialogProps = {
             id: UNIVER_SHEET_PERMISSION_ALERT_DIALOG_ID,
             title: { title: '' },
@@ -327,7 +302,7 @@ export class SheetPermissionInterceptorRenderController extends RxDisposable imp
         }
 
         if (!permission) {
-            this._haveNotPermissionHandle(errorMsg);
+            this.haveNotPermissionHandle(errorMsg);
         }
     };
 
@@ -689,318 +664,42 @@ export class SheetPermissionInterceptorRenderController extends RxDisposable imp
         return this._permissionCheckWithRanges(permissionTypes, ranges, unitId, subUnitId);
     }
 
-    private _initHeaderMovePermissionInterceptor() {
-        this._headerMoveRenderController.interceptor.intercept(this._headerMoveRenderController.interceptor.getInterceptPoints().HEADER_MOVE_PERMISSION_CHECK, {
-            handler: (defaultValue: Nullable<boolean>, selectionRange: IRange) => {
-                const target = getSheetCommandTarget(this._univerInstanceService);
-                if (!target) {
-                    return false;
+    private _permissionCheckWithFormula(params: ISetRangeValuesCommandParams) {
+        const value = params.value as ICellData;
+        const range = params.range;
+        const formulaString = value.f;
+        if (formulaString) {
+            const definedNameStr = formulaString.substring(1);
+            const workbook = this._univerInstanceService.getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
+            const unitId = params.unitId ?? workbook.getUnitId();
+            const definedName = this._definedNamesService.getValueByName(unitId, definedNameStr);
+            if (definedName) {
+                let formulaOrRefString = definedName.formulaOrRefString;
+                if (formulaOrRefString.startsWith(operatorToken.EQUALS)) {
+                    formulaOrRefString = formulaOrRefString.slice(1);
                 }
-                const { worksheet, unitId, subUnitId } = target;
-
-                const workSheetEditPermission = this._permissionService.getPermissionPoint(new WorksheetEditPermission(unitId, subUnitId).id)?.value ?? false;
-                if (!workSheetEditPermission) {
-                    return false;
-                }
-
-                if (!selectionRange) {
-                    return true;
-                }
-
-                const protectionLapRange = this._rangeProtectionRuleModel.getSubunitRuleList(unitId, subUnitId).reduce((p, c) => {
-                    return [...p, ...c.ranges];
-                }, [] as IRange[]).filter((range) => {
-                    return Rectangle.intersects(range, selectionRange);
-                });
-
-                const haveNotPermission = protectionLapRange.some((range) => {
-                    const { startRow, startColumn, endRow, endColumn } = range;
-                    for (let row = startRow; row <= endRow; row++) {
-                        for (let col = startColumn; col <= endColumn; col++) {
-                            const permission = (worksheet.getCell(row, col) as (ICellDataForSheetInterceptor & { selectionProtection: ICellPermission[] }))?.selectionProtection?.[0];
-                            if (permission?.[UnitAction.Edit] === false) {
-                                return true;
-                            }
-                        }
-                    }
-                    return false;
-                });
-
-                return !haveNotPermission;
-            },
-        });
-    }
-
-    private _initHeaderResizePermissionInterceptor() {
-        this._headerResizeRenderController.interceptor.intercept(this._headerResizeRenderController.interceptor.getInterceptPoints().HEADER_RESIZE_PERMISSION_CHECK, {
-            handler: (defaultValue: Nullable<boolean>, rangeParams: { row?: number; col?: number }) => {
-                const target = getSheetCommandTarget(this._univerInstanceService);
-                if (!target) {
-                    return false;
-                }
-                const { worksheet, unitId, subUnitId } = target;
-
-                const workSheetEditPermission = this._permissionService.getPermissionPoint(new WorksheetEditPermission(unitId, subUnitId).id)?.value ?? false;
-                if (!workSheetEditPermission) {
-                    return false;
-                }
-
-                if (rangeParams.row) {
-                    const setRowStylePermission = this._permissionService.getPermissionPoint(new WorksheetSetRowStylePermission(unitId, subUnitId).id)?.value ?? false;
-                    if (setRowStylePermission === false) {
-                        return false;
-                    }
-                } else if (rangeParams.col) {
-                    const setColStylePermission = this._permissionService.getPermissionPoint(new WorksheetSetColumnStylePermission(unitId, subUnitId).id)?.value ?? false;
-                    if (setColStylePermission === false) {
-                        return false;
-                    }
-                }
-
-                let selectionRange: Nullable<IRange>;
-
-                if (rangeParams.row !== undefined) {
-                    selectionRange = {
-                        startRow: rangeParams.row,
-                        endRow: rangeParams.row,
-                        startColumn: 0,
-                        endColumn: worksheet.getColumnCount() - 1,
-                    };
-                } else if (rangeParams.col !== undefined) {
-                    selectionRange = {
-                        startRow: 0,
-                        endRow: worksheet.getRowCount() - 1,
-                        startColumn: rangeParams.col,
-                        endColumn: rangeParams.col,
-                    };
-                }
-
-                if (!selectionRange) {
-                    return true;
-                }
-
-                const protectionLapRange = this._rangeProtectionRuleModel.getSubunitRuleList(unitId, subUnitId).reduce((p, c) => {
-                    return [...p, ...c.ranges];
-                }, [] as IRange[]).filter((range) => {
-                    return Rectangle.intersects(range, selectionRange);
-                });
-
-                const haveNotPermission = protectionLapRange.some((range) => {
-                    const { startRow, startColumn, endRow, endColumn } = range;
-                    for (let row = startRow; row <= endRow; row++) {
-                        for (let col = startColumn; col <= endColumn; col++) {
-                            const permission = (worksheet.getCell(row, col) as (ICellDataForSheetInterceptor & { selectionProtection: ICellPermission[] }))?.selectionProtection?.[0];
-                            if (permission?.[UnitAction.Edit] === false) {
-                                return true;
-                            }
-                        }
-                    }
-                    return false;
-                });
-
-                return !haveNotPermission;
-            },
-        });
-    }
-
-    private _initRangeFillPermissionInterceptor() {
-        this._selectionRenderService.interceptor.intercept(this._selectionRenderService.interceptor.getInterceptPoints().RANGE_FILL_PERMISSION_CHECK, {
-            handler: (_: Nullable<boolean>, position: { x: number; y: number; skeleton: SpreadsheetSkeleton }) => {
-                const target = getSheetCommandTarget(this._univerInstanceService);
-                if (!target) {
-                    return false;
-                }
-                const { worksheet, unitId, subUnitId } = target;
-
-                const workSheetEditPermission = this._permissionService.getPermissionPoint(new WorksheetEditPermission(unitId, subUnitId).id)?.value ?? false;
-                if (!workSheetEditPermission) {
-                    return false;
-                }
-
-                const ranges = this._selectionManagerService.getSelections()?.map((selection) => {
-                    return selection.range;
-                });
-
-                const selectionRange = ranges?.find((range) => {
-                    const cellPosition = position.skeleton.getCellByIndex(range.endRow, range.endColumn);
-                    const missX = Math.abs(cellPosition.endX - position.x);
-                    const missY = Math.abs(cellPosition.endY - position.y);
-                    return missX <= 5 && missY <= 5;
-                });
-
-                if (!selectionRange) {
-                    return true;
-                }
-
-                const { startRow, endRow, startColumn, endColumn } = selectionRange;
-
-                for (let row = startRow; row <= endRow; row++) {
-                    for (let col = startColumn; col <= endColumn; col++) {
-                        const permission = (worksheet.getCell(row, col) as (ICellDataForSheetInterceptor & { selectionProtection: ICellPermission[] }))?.selectionProtection?.[0];
-                        if (permission?.[UnitAction.Edit] === false || permission?.[UnitAction.View] === false) {
+                const refRangesArr = formulaOrRefString.split(',');
+                for (let i = 0; i < refRangesArr.length; i++) {
+                    const refRange = refRangesArr[i];
+                    const sequenceGrid = deserializeRangeWithSheet(refRange);
+                    if (sequenceGrid.sheetName) {
+                        const targetSheet = workbook.getSheetBySheetName(sequenceGrid.sheetName);
+                        if (!targetSheet) {
                             return false;
+                        }
+                        const { startRow, endRow, startColumn, endColumn } = sequenceGrid.range;
+                        for (let i = startRow; i <= endRow; i++) {
+                            for (let j = startColumn; j <= endColumn; j++) {
+                                const permission = (targetSheet.getCell(i, j) as (ICellDataForSheetInterceptor & { selectionProtection: ICellPermission[] }))?.selectionProtection?.[0];
+                                if (permission?.[UnitAction.View] === false) {
+                                    return false;
+                                }
+                            }
                         }
                     }
                 }
                 return true;
-            },
-        });
-    }
-
-    private _initRangeMovePermissionInterceptor() {
-        this._selectionRenderService.interceptor.intercept(this._selectionRenderService.interceptor.getInterceptPoints().RANGE_MOVE_PERMISSION_CHECK, {
-            handler: (_: Nullable<boolean>, _cellInfo: null) => {
-                const target = getSheetCommandTarget(this._univerInstanceService);
-                if (!target) {
-                    return false;
-                }
-                const { worksheet, unitId, subUnitId } = target;
-
-                const workSheetEditPermission = this._permissionService.getPermissionPoint(new WorksheetEditPermission(unitId, subUnitId).id)?.value ?? false;
-                if (!workSheetEditPermission) {
-                    return false;
-                }
-
-                const ranges = this._selectionManagerService.getSelections()?.map((selection) => {
-                    return selection.range;
-                });
-
-                const ruleRanges = this._rangeProtectionRuleModel.getSubunitRuleList(unitId, subUnitId).reduce((p, c) => {
-                    return [...p, ...c.ranges];
-                }, [] as IRange[]);
-
-                const permissionLapRanges = ranges?.filter((range) => {
-                    return ruleRanges.some((ruleRange) => {
-                        return Rectangle.intersects(ruleRange, range);
-                    });
-                });
-
-                const haveNotPermission = permissionLapRanges?.some((range) => {
-                    const { startRow, startColumn, endRow, endColumn } = range;
-                    for (let row = startRow; row <= endRow; row++) {
-                        for (let col = startColumn; col <= endColumn; col++) {
-                            const permission = (worksheet.getCell(row, col) as (ICellDataForSheetInterceptor & { selectionProtection: ICellPermission[] }))?.selectionProtection?.[0];
-                            if (permission?.[UnitAction.Edit] === false) {
-                                return true;
-                            }
-                        }
-                    }
-                    return false;
-                });
-
-                return !haveNotPermission;
-            },
-        });
-    }
-
-    private _initSheetFindPermissionInterceptor() {
-        this._sheetsFindReplaceController.interceptor.intercept(this._sheetsFindReplaceController.interceptor.getInterceptPoints().FIND_PERMISSION_CHECK, {
-            handler: (_: Nullable<boolean>, _cellInfo: { row: number; col: number; unitId: string; subUnitId: string }) => {
-                const { row, col, subUnitId } = _cellInfo;
-                const workbook = this._univerInstanceService.getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
-                const worksheet = workbook.getSheetBySheetId(subUnitId);
-                if (!worksheet) {
-                    return false;
-                }
-                const permission = (worksheet.getCell(row, col) as (ICellDataForSheetInterceptor & { selectionProtection: ICellPermission[] }))?.selectionProtection?.[0];
-                return permission?.[UnitAction.View] !== false;
-            },
-        });
-    }
-
-    private _initDataValidationPermissionInterceptor() {
-        this._dataValidationController.interceptor.intercept(this._dataValidationController.interceptor.getInterceptPoints().DATA_VALIDATION_PERMISSION_CHECK, {
-            handler: (_: Nullable<(ISheetDataValidationRule & { disable?: boolean })[]>, rules: ISheetDataValidationRule[]) => {
-                const workbook = this._univerInstanceService.getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
-                const worksheet = workbook.getActiveSheet();
-                const rulesByPermissionCheck = rules.map((rule) => {
-                    const { ranges } = rule;
-                    const haveNotPermission = ranges?.some((range) => {
-                        const { startRow, startColumn, endRow, endColumn } = range;
-                        for (let row = startRow; row <= endRow; row++) {
-                            for (let col = startColumn; col <= endColumn; col++) {
-                                const permission = (worksheet.getCell(row, col) as (ICellDataForSheetInterceptor & { selectionProtection: ICellPermission[] }))?.selectionProtection?.[0];
-                                if (permission?.[UnitAction.Edit] === false || permission?.[UnitAction.View] === false) {
-                                    return true;
-                                }
-                            }
-                        }
-                        return false;
-                    });
-                    if (haveNotPermission) {
-                        return { ...rule, disable: true };
-                    } else {
-                        return { ...rule };
-                    }
-                });
-                return rulesByPermissionCheck;
-            },
-        });
-    }
-
-    private _initConditionalFormattingPermissionInterceptor() {
-        this._conditionalFormattingClearController.interceptor.intercept(this._conditionalFormattingClearController.interceptor.getInterceptPoints().CONDITIONAL_FORMATTING_PERMISSION_CHECK, {
-            handler: (_: Nullable<(IConditionFormattingRule<IConditionalFormattingRuleConfig> & { disable?: boolean })[]>, rules: ((IConditionFormattingRule<IConditionalFormattingRuleConfig> & { disable?: boolean }))[]) => {
-                const workbook = this._univerInstanceService.getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
-                const worksheet = workbook.getActiveSheet();
-                const rulesByPermissionCheck = rules.map((rule) => {
-                    const ranges = rule.ranges;
-                    const haveNotPermission = ranges?.some((range) => {
-                        const { startRow, startColumn, endRow, endColumn } = range;
-                        for (let row = startRow; row <= endRow; row++) {
-                            for (let col = startColumn; col <= endColumn; col++) {
-                                const permission = (worksheet.getCell(row, col) as (ICellDataForSheetInterceptor & { selectionProtection: ICellPermission[] }))?.selectionProtection?.[0];
-                                if (permission?.[UnitAction.Edit] === false || permission?.[UnitAction.View] === false) {
-                                    return true;
-                                }
-                            }
-                        }
-                        return false;
-                    });
-                    if (haveNotPermission) {
-                        return { ...rule, disable: true };
-                    } else {
-                        return { ...rule };
-                    }
-                });
-                return rulesByPermissionCheck;
-            },
-        });
-    }
-
-    private _initStatusBarPermissionInterceptor() {
-        this.disposeWithMe(
-            this._statusBarController.interceptor.intercept(this._statusBarController.interceptor.getInterceptPoints().STATUS_BAR_PERMISSION_CORRECT, {
-                priority: 100,
-                handler: (defaultValue, originValue) => {
-                    const target = getSheetCommandTarget(this._univerInstanceService);
-                    if (!target) {
-                        return defaultValue ?? [];
-                    }
-                    const { worksheet } = target;
-                    originValue.forEach((item) => {
-                        const itemValue = item.getArrayValue();
-                        const startRow = item.getCurrentRow();
-                        const startCol = item.getCurrentColumn();
-                        itemValue.forEach((row, rowIndex) => {
-                            row.forEach((col, colIndex) => {
-                                const permission = (worksheet.getCell(rowIndex + startRow, colIndex + startCol) as (ICellDataForSheetInterceptor & { selectionProtection: ICellPermission[] }))?.selectionProtection?.[0];
-                                if (permission?.[UnitAction.View] === false) {
-                                    itemValue[rowIndex][colIndex] = NullValueObject.create();
-                                }
-                            });
-                        });
-                    });
-                    return originValue;
-                },
-            })
-        );
-    }
-
-    private _initDvFormulaPermissionInterceptor() {
-        this._dataValidationFormulaController.interceptor.intercept(this._dataValidationFormulaController.interceptor.getInterceptPoints().DV_FORMULA_PERMISSION_CHECK, {
-            priority: 100,
-            handler: (_, formulaString: string) => {
+            } else {
                 const sequenceNodes = this._lexerTreeBuilder.sequenceNodesBuilder(formulaString);
                 if (!sequenceNodes) {
                     return true;
@@ -1034,139 +733,6 @@ export class SheetPermissionInterceptorRenderController extends RxDisposable imp
                             if (permission?.[UnitAction.View] === false) {
                                 return false;
                             }
-                        }
-                    }
-                }
-                return true;
-            },
-        });
-    }
-
-    private _initFreezePermissionInterceptor() {
-        this._headerFreezeRenderController.interceptor.intercept(this._headerFreezeRenderController.interceptor.getInterceptPoints().FREEZE_PERMISSION_CHECK, {
-            handler: (_: Nullable<boolean>, __) => {
-                const workbook = this._univerInstanceService.getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
-                const worksheet = workbook.getActiveSheet();
-                if (!worksheet) {
-                    return false;
-                }
-                const permission = this._permissionService.getPermissionPoint(new WorkbookEditablePermission(workbook.getUnitId()).id)?.value ?? false;
-                return permission;
-            },
-        });
-    }
-
-    private _initFormulaEditorPermissionInterceptor() {
-        this._univerInstanceService.unitAdded$.pipe(filter((unitInstance) => unitInstance.type === UniverInstanceType.UNIVER_DOC), first()).subscribe(() => {
-            const formulaEditorController = this._injector.get(FormulaEditorController);
-            formulaEditorController.interceptor.intercept(formulaEditorController.interceptor.getInterceptPoints().FORMULA_EDIT_PERMISSION_CHECK, {
-                handler: (_: Nullable<boolean>, cellInfo: { row: number; col: number }) => {
-                    const target = getSheetCommandTarget(this._univerInstanceService);
-                    if (!target) {
-                        return false;
-                    }
-                    const { unitId, subUnitId } = target;
-
-                    const worksheetViewPermission = this._permissionService.getPermissionPoint(new WorksheetEditPermission(unitId, subUnitId).id)?.value ?? false;
-                    if (!worksheetViewPermission) {
-                        return false;
-                    }
-
-                    const { row, col } = cellInfo;
-                    const permissionList = this._rangeProtectionRuleModel.getSubunitRuleList(unitId, subUnitId).filter((rule) => {
-                        return rule.ranges.some((range) => {
-                            return Rectangle.intersects(range, { startRow: row, endRow: row, startColumn: col, endColumn: col });
-                        });
-                    });
-
-                    const permissionIds = permissionList.map((rule) => new RangeProtectionPermissionViewPoint(unitId, subUnitId, rule.permissionId).id);
-                    const rangeViewPermission = this._permissionService.composePermission(permissionIds).every((permission) => permission.value);
-                    return rangeViewPermission;
-                },
-            });
-        });
-    }
-
-    private _initClipboardHook() {
-        this.disposeWithMe(
-            this._sheetClipboardService.addClipboardHook({
-                id: SHEET_PERMISSION_PASTE_PLUGIN,
-                onBeforePaste: (pasteTo) => {
-                    const [ranges] = virtualizeDiscreteRanges([pasteTo.range]).ranges;
-                    const startRange = this._selectionManagerService.getLast()?.range;
-                    if (!startRange) {
-                        return false;
-                    }
-                    const targetRange = {
-                        startRow: startRange.startRow + ranges.startRow,
-                        endRow: startRange.startRow + ranges.endRow,
-                        startColumn: startRange.startColumn + ranges.startColumn,
-                        endColumn: startRange.startColumn + ranges.endColumn,
-                    };
-
-                    const workbook = this._univerInstanceService.getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
-                    const worksheet = workbook.getActiveSheet();
-                    const { startRow, endRow, startColumn, endColumn } = targetRange;
-
-                    let hasPermission = true;
-
-                    for (let row = startRow; row <= endRow; row++) {
-                        for (let col = startColumn; col <= endColumn; col++) {
-                            const permission = (worksheet.getCell(row, col) as (ICellDataForSheetInterceptor & { selectionProtection: ICellPermission[] }))?.selectionProtection?.[0];
-                            if (permission?.[UnitAction.Edit] === false) {
-                                hasPermission = false;
-                                break;
-                            }
-                        }
-                    }
-
-                    if (!hasPermission) {
-                        this._haveNotPermissionHandle(this._localService.t('permission.dialog.pasteErr'));
-                    }
-
-                    return hasPermission;
-                },
-            })
-        );
-    }
-
-    private _permissionCheckWithFormula(params: ISetRangeValuesCommandParams) {
-        const value = params.value as ICellData;
-        const range = params.range;
-        const formulaString = value.f;
-        if (formulaString) {
-            const sequenceNodes = this._lexerTreeBuilder.sequenceNodesBuilder(formulaString);
-            if (!sequenceNodes) {
-                return true;
-            }
-            for (let i = 0; i < sequenceNodes.length; i++) {
-                const node = sequenceNodes[i];
-                if (typeof node === 'string') {
-                    continue;
-                }
-                const { token } = node;
-                const sequenceGrid = deserializeRangeWithSheet(token);
-                const workbook = this._univerInstanceService.getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
-                let targetSheet: Nullable<Worksheet> = workbook.getActiveSheet();
-                const unitId = workbook.getUnitId();
-                if (sequenceGrid.sheetName) {
-                    targetSheet = workbook.getSheetBySheetName(sequenceGrid.sheetName);
-                    if (!targetSheet) {
-                        return false;
-                    }
-                    const subUnitId = targetSheet?.getSheetId();
-                    const viewPermission = this._permissionService.getPermissionPoint(new WorksheetViewPermission(unitId, subUnitId).id);
-                    if (!viewPermission) return false;
-                }
-                if (!targetSheet) {
-                    return false;
-                }
-                const { startRow, endRow, startColumn, endColumn } = sequenceGrid.range;
-                for (let i = startRow; i <= endRow; i++) {
-                    for (let j = startColumn; j <= endColumn; j++) {
-                        const permission = (targetSheet.getCell(i, j) as (ICellDataForSheetInterceptor & { selectionProtection: ICellPermission[] }))?.selectionProtection?.[0];
-                        if (permission?.[UnitAction.View] === false) {
-                            return false;
                         }
                     }
                 }
